+++ CHANGELOG.md
@@ -23,6 +23,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-1497 Change all buttons to a flat design
 - YLP-1500 Patient can join a team from header
 - YLP-1511 Display message date, status and the number of unread messages in chat widget
+- YLP-1518 Implement drawer filters
 ### Engineering Use
 - YLP-1256 Split Profile component into smaller subcomponents
 - YLP-1319 Switch-roles dialogs CSS not consistent with other dialogs